Adds clarification about uploading ssh key
The paragraph describing SSH key upload does not make it clear that the
upload is being done on review.openstack.org as opposed to the Launchpad
account.  If the user of the manual does not realize the link is
for review.openstack.org the user may never add their SSH key to Gerrit.

Change-Id: Ied449976ee335a82093d1231005f06482f7be1df