Creates a directory for non-indexed content
Right now we don't have a place for files that are not guides by
themselves but are content to be linked to other guides.

This patch creates a directory called content for files not linked
to the index.rst file. It also includes a file for instructions
for using the developer sandbox that can be linked from
developer.rst.

Change-Id: I2cf79758fd088d3eb294d8828dc87c468f47dbea