Add details on using git-review over https
Add a small section documenting how to use git-review over an
https connection. The information had been added to the wiki
but the whole page was removed and points to the current docs
which were lacking this information.

Added changes suggested during review, including down-grading
the header one level and clarifying that the https configuration
is optional and only needed when internet access is restricted

Change-Id: Iddd9f5e5b1f37b9969fa0aa8b30b3cfe9c83cb5f
Closes-Bug: #1427965