Reverse dependency relationship in project creation
This change depends on a governance change that makes it explicit
that the TC does not need to prior-approve additions of git repos
to existing OpenStack projects, therefore, change the instructions
to specify that the project should be created first, but a change
prepared for the governance repo to add it to the list.

However, if an entirely new project is being added to OpenStack,
then the dependency relationship should be creation depends-on
addition to governance repo.

Change-Id: I898941f16a1b75fa5f983e19e5683cb06e83c082
Depends-On: I34acfe4ffe5a0af6a941e70a5315a07d036e6c24