Initial infra manual for infra-cloud
This does not include many details, and for good reason. Many of these
things are open questions. I am not sure if we should add a section on
the specific hardware/locations or not, so that is up to date.

The document is written in the present tense, so as we make these things
true, they won't have to be altered. The status section should reflect
how close to reality the documentation is until we implement some of it.

Change-Id: Icb35adf70fa98e64dbbe0464b95af3b5de51a980