developer's guide: Add bug handling description
The bug handling in OpenStack is described over a few wiki pages [1][2]
[3][4]. It also missed some information a new developer would need to
understand the process. An adoption and review of the process with the
core reviewers and the bug-supervisors is hard to do. After this patch
merges, the mentioned wiki pages should be trimmed down and point to
this manual entry.

This patch set adds content to the developer's guide and should be the
"single source of truth" one needs to read to be able to participate in
the bug handling process of OpenStack.

[1] https://wiki.openstack.org/wiki/Bugs
[2] https://wiki.openstack.org/wiki/BugTriage
[3] https://wiki.openstack.org/wiki/BugTags
[4] https://wiki.openstack.org/wiki/Nova/BugTriage

Change-Id: Ia699fd1965b475f2633ecfc555bc766eb5aa3404