developer's guide: Add bug handling description
The bug handling in OpenStack is described over a few wiki pages. It
also missed some information a new developer would need to understand
the process. An adoption and review of the process with the core
reviewers and the bug-supervisors is hard to do.

This patch set adds content to the developer's guide and should be the
"single source of truth" one needs to read to be able to participate in
the bug handling process of OpenStack.

Change-Id: Ia699fd1965b475f2633ecfc555bc766eb5aa3404