Change many uses of repository to project
This is a manual that describes how to use the OpenStack project
infrastructure.  As the TC changes its mind on what words to use
for what things, and redefines words to mean new things, this
manual does not need to be updated to match.  That the TC happens
to use the word "project" to mean something in organizing its
governance structure one day, but not the next, has little
relation to how our tools are configured.

Gerrit has an understanding of "projects".  So does "zuul".  It
will be better for us to use the terminology we have
traditionally used and that many of our tools and users use.  We
can certainly use the term "repository" when there might
legitimately be confusion.  But let us not force the manual into
incomprehensibility by blindly adopting TC terminology where it
is not needed.

This is a partial revert of
I574ab8e2c94eadce76cff9f7ea561dfa5d28b89c.  However, that change
did add many useful clarifications, so it is not a full revert.

In short, I believe that this manual should use the common term
"project" in almost all cases.  The term "project" refers to the
overall idea that there is a bunch of people working on a bunch
of code/text/etc.  It can also refer to that actual collection of
code/text/etc (for instance, a project can be bundled up into a
tarball, and extracted into a directory).  When a tool interacts
with that collection of code/text/etc, it interacts with the
project (even if it does so via the mechanism of git).  There are
times when one needs to refer to the actual source code
management system of the project, that is, "git", and the actual
technical implementations of that SCM.  In those cases where it
is important to distinguish the actual attributes of the SCM from
the project, it is useful to use the word "repository".

Change-Id: I7cc37c32841fb33b38c3de90f426595de24a31d6