Move some content to HOWTOs
The primary guides should be reference documentation and generally
should and are sufficient to stand on their own.  However, we have
added some step-by-step instructions for some tasks that don't fit
that idea, yet are very useful (and often more likely to be read
and followed than the actual documentation because of their more
limited scope).  Reorganize the index page to move them away from
the three primary guides and give them their own pride of place
in a "HOWTOs" section.

Change-Id: I3c7d6b40f3c3909b8cb439af6a0a29a6ff781582