Improve docs on setting up development environment
We need newer than distro pip's in tox virtualenvs, which means we
need a new virtualenv.

Further, because some platforms mangle pip by default we can't use
the system pip, as the installed tools won't be system installed -
and often incorrect python paths end up making pip installed things
that shadow system packages behave incorrectly.

So instead use get-pip.py and remove the things we're going to use in
getting the virtualenvs up and running.

An alternative approach that doesn't require altering system installed
packages is documented on my blog [1] but I'm not sure if it's
sufficiently low friction to go in the dev docs.

[1] https://rbtcollins.wordpress.com/2015/07/12/bootstrapping-developer-environments-for-openstack/

Co-Authored-By: Samuel de Medeiros Queiroz <samueldmq@gmail.com>

Change-Id: I697af934aaca002da0be3555d1e1829099ead4b3