Enhance zuul setup section
Remove integrated-gate since that is very specific and not used in
general anymore.

Add a new paragraph explaining that each job needs to be defined in
jenkins first so that it can be triggered by zuul. This should avoid
some confusion where people only touch zuul setup without going through
the whole file.

Change-Id: Iabdadd0ef053bdaa5e10d1be95e7c7025d22ebb2