Cookiecutter fails with "directory already exists"
Creating the initial commit for a reposoitory using
cookiecutter along the project creators guide fails with "directory
already exists". That happens, as cookiecutter tries to create
a new directory for the project, but it already exists, as the creators
guide asked to check it out one step before. Using the -f option works
perfectly here. The initial commit will be put on top of the already
existing .gitreview commit of the existing project.

Change-Id: I11ac237a551a83c2570475a248855805be989b2f