Add small note to check CI status when discovering failures
The automated testing section is the one that is linked to in the
comment left on a review.  It does not make any mention of problems
in the infrastructure.

This patch addresses this and adds a simple note to left developers
know where to look for if there are any issues.

Change-Id: I3f8a5d68b552eb71fb758af4e860e5cba2e46c26