Add some information about OpenStack IRC usage
IRC is an essential part of being an OpenStack developer. There are a
billion ways to interact with IRC, and the best way is to learn
about them. However, some people are starting completely from scratch,
so collecting some information for them with pointers to full docs is
maybe a nice thing to do.

Change-Id: Ie4e3ec74eaac6f8ae22f9af7d2c7b32e2fc280f9