promote IRC channel creation info into its own section
Previously the link about IRC channel creation was a bit hard to find,
because it was not obviously discoverable from the TOC, so promote it
to its own section.

Also make it come before the information about setting up GerritBot
within a channel, not after, since the logical order is that you need
to set up a channel before you can configure bots within in.

Change-Id: If6dbae7e9b0c17e0bc45e710bf9eea07110c4215