Update the PyPI registration docs
python setup.py register and twine register are no longer supported.
PyPI now requires a package by uploaded as the registration mechanism.

Update the instructions to describe pushing a 0.0.0.0a1 tag for initial
registration. Also, make the tagging instructions in the creators guide
be specifically about people making new unofficial projects and point
people to the tagging section of the drivers guide and the
openstack/releases README for the other cases.

Change-Id: I94f85b67646ceccea8d09b0a576291aad261e911