Clarify the point of a repo in a zuulv3 job name
There were a couple of places in the zuulv3 job name
guidelines where it said to not use the repository name
in the job name, but then at the end it says to prefix
the job name with the repo if the job is defined in a
specific repo.

This clarifies that point and also fixes some grammar
and typo issues, plus adds another example.

This came out of some review discussion on nova change
I41b03a34795efe139d5911c605cdbd3c47a2f059.

Change-Id: I52b9c7db145d4322c1084337c89e49d0c2d01c6e