Correct django template pattern in babel-django.cfg
The current recommended file pattern "templates/**.html" does not
match Django templates in subdirectories.
This needs to be "**/templates/**.html".

This was fixed in horizon two years ago in Liberty, but somehow
the incorrect file pattern is recommended in the infra manual
and used in several number of horizon plugins.
It is worth fixed in the infra manual.

Change-Id: I4a06faaca8726b386369b9066324f1625deb3507
Story: #2001514
Task: #6314