update retirement instructions to keep the .gitreview file
The release scripts that try to clone repositories look for .gitreview
because they use "git review -s" to prepare the local repo to propose
patches, create tags, or make branches. If the .gitreview file is not
present on master the clone operation fails. This patch updates the
retirement instructions to ask folks to keep .gitreview in place to
ensure smooth operation in the automation.

Change-Id: Ie6f94db0306a6c603a642fde361f7cdf3ea7e5ab
Signed-off-by: Doug Hellmann <doug@doughellmann.com>