Push PTI job config into projects
The Infra team would like to avoid needing to review every PTI job
related update to openstack projects. Make it clear that it is ok for
projects to manage this config in repo so that they can manage and
review it themselves. This takes advantage of the new shiny features
that Zuul v3 provides in a way that should make both infra and
individual projects happier.

Change-Id: I51bd6c5bbcb8135c3ea4fed6941613643a3bcead