Update .gitreview wording in retire process
We had a note in the repo retirement process regarding steps to take
before removing the .gitreview file. Later we determined that it is best
to keep that file in the repo rather than removing it, but the note
mentioning its removal was kept after other wording updates noting it
should stay.

This removes the removal reference and updates the notes about keeping
the file to attempt to make things more clear.

Change-Id: Id4ee69de31569645c3486f7b8508db3d6e64da8f
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>