translation: drop babel extractor definitions
babel extractors are now registered via python entry points,
so there is no need to declare babel extractors in babel configs.

This allows us to switch the message extractor easily.
For example, the horizon team plans to switch the babel extractor
from django-babel to enmerkar (a fork of django-babel) near future
as django-babel does not work with Django 2.2 (the recent LTS of Django)
and looks unmaintained for over two years.

Change-Id: I4580aa4b0f1ba2f6fd4dcaa3510b665ca5a77024