Improve job naming convention docs
We are now well past the Zuulv2 -> v3 transition and the old v2 context
is likely not very useful for many of our users. Improve the job naming
convention docs to remove transition specific suggestions and instead
simply provide a set of rules that work for our current Zuul deployment.

Change-Id: Ifc069d366cd7fd27c4bbc2b3331c78a6477659a1