OpenDev Update for Creator's Guide
This is a first iteration on making the Creator's Guide ready for
OpenDev. The whole file needs some more sections moved out, let's tackle
those separately.

It removes the section "Updating devstack-vm-gate-wrap.sh" since that is
not needed anymore.

Remove section "Add Project to the Requirements List" since the syncing
is not done anymore.

Change-Id: Idd949b3459cb136391abea8039d44ee80e8b1992